Now you can chose to display the timestamps in local time of the client
If you want to show times in UTC set useClientTimezone in config.js = 0 else
the timestamps and headers in the tables will be displayed in the local timezone
of the client-system. So if you are showing it in Japan with your local time there
you will get other results than showing it in New York.